Function readADC() not returns an 'int' instead of 'uint8'
Previously, the value read was unstable, and the root cause was that the actual value read from ADC was longer than 8 bits. Changing it to 'int' should solve this issue. Calibration of the sensor measurements still needs to be done in software though.

nuttx.tftf added to the folder and is available at nuttx/nuttx/